vfio-hook: Improve rebinding logic
Currently we write the device to rebind's vendor and device to vfio-pci's
new_id file.  That could cause vfio-pci to attach to devices other than the
one we intend which have the same vendor:device.  That might not be correct
in all circumstances.

To avoid this, we can use the device's "driver_override" file instead which
overrides the ID matching logic just for this specific device.  It also
avoids a bit of fiddling around to retrieve the vendor and id, which we
otherwise no longer care about.

While we're there, make some assorted cleanups to the device rebinding
path.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>